Rename middleware directory
