+++ apps/api/src/lib/middleware/authenticate.ts
