+++ apps/api/src/lib/middleware/capture-errors.ts
