+++ apps/api/src/lib/middleware/default-handler.ts
