+++ apps/api/src/lib/middleware/default-responder.ts
