+++ apps/api/src/pages/api/git/review-comments/_get.ts
@@ -3,7 +3,7 @@ import { querySchema } from "@floe/requests/git/review-comments/_get";
 import type { GetGitReviewCommentsResponse } from "@floe/requests/git/review-comments/_get";
 import type { NextApiRequestExtension } from "~/types/middleware";
 import { getOctokit } from "~/lib/github/octokit";
-import { defaultResponder } from "~/lib/helpers/default-responder";
+import { defaultResponder } from "~/lib/middleware/default-responder";
 import { zParse } from "~/utils/z-parse";
 
 async function handler({
