+++ apps/api/src/pages/api/git/review-comments/_post.ts
@@ -3,7 +3,7 @@ import { querySchema } from "@floe/requests/git/review-comments/_post";
 import type { PostGitReviewCommentsResponse } from "@floe/requests/git/review-comments/_post";
 import type { NextApiRequestExtension } from "~/types/middleware";
 import { getOctokit } from "~/lib/github/octokit";
-import { defaultResponder } from "~/lib/helpers/default-responder";
+import { defaultResponder } from "~/lib/middleware/default-responder";
 import { zParse } from "~/utils/z-parse";
 
 async function handler({
